#75 Extract item data: removed GameItemList completely
